refactor: lint dirty modules plugin (#59)
* refactor: use filter on filetimestamps' keys

as opposed to reduce. also use lodash.defaultto for comparing numeric
values to avoid falsy JS bugs

* refactor: use Arrange Act Assert model for tests

and clean up some assertions and specs

* refactor: pulls out defaultFilesGlob constant

* refactor: clean up typos

* refactor: use .test.js file suffix

to enable the recursive flag in mocha.

* refactor: better names for all fixtures

* tests: add coverage for without plugin config

* tests: update NoErrorsPlugin tests

Verify that existing behaviour works as expected

* refactor: minor test tweaks

* tests: correct test of empty plugin config

* tests: remove extra context from lint dirty modules plugin
