+++ lib/constants.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  defaultFilesGlob: '**/*.s?(c|a)ss'
+};
