+++ lib/lint-dirty-modules-plugin.js
@@ -1,7 +1,8 @@
 'use strict';
-var minimatch = require('minimatch');
-var reduce = require('lodash.reduce');
+
 var assign = require('object-assign');
+var defaultTo = require('lodash.defaultto');
+var minimatch = require('minimatch');
 var runCompilation = require('./run-compilation');
 
 /**
@@ -9,7 +10,6 @@ var runCompilation = require('./run-compilation');
  *
  * @param compiler - webpack compiler object
  * @param options - stylelint nodejs options
- * @param callback <function(options, compilitaion)> - callback to call on emit
  */
 function LintDirtyModulesPlugin (compiler, options) {
   this.startTime = Date.now();
@@ -17,9 +17,9 @@ function LintDirtyModulesPlugin (compiler, options) {
   this.isFirstRun = true;
   this.compiler = compiler;
   this.options = options;
-  compiler.plugin('emit',
-    this.lint.bind(this) // bind(this) is here to prevent context overriding by webpack
-  );
+
+  // bind(this) is here to prevent context overriding by webpack
+  compiler.plugin('emit', this.lint.bind(this));
 }
 
 /**
@@ -31,16 +31,18 @@ function LintDirtyModulesPlugin (compiler, options) {
  * @param callback - to be called when execution is done
  * @returns {*}
  */
-LintDirtyModulesPlugin.prototype.lint = function (compilation, callback) {
+LintDirtyModulesPlugin.prototype.lint = function lint (compilation, callback) {
   if (this.isFirstRun) {
     this.isFirstRun = false;
     this.prevTimestamps = compilation.fileTimestamps;
     return callback();
   }
+
   var dirtyOptions = assign({}, this.options);
   var glob = dirtyOptions.files.join('|');
   var changedFiles = this.getChangedFiles(compilation.fileTimestamps, glob);
   this.prevTimestamps = compilation.fileTimestamps;
+
   if (changedFiles.length) {
     dirtyOptions.files = changedFiles;
     runCompilation.call(this, dirtyOptions, this.compiler, callback);
@@ -57,18 +59,16 @@ LintDirtyModulesPlugin.prototype.lint = function (compilation, callback) {
  * @param fileTimestamps - an object with keys as filenames and values as their timestamps.
  * e.g. {'/filename.scss': 12444222000}
  * @param glob - glob pattern to match files
+ * @returns {Array} list of globs that contain changed files
  */
-LintDirtyModulesPlugin.prototype.getChangedFiles = function (fileTimestamps, glob) {
-  return reduce(fileTimestamps, function (changedStyleFiles, timestamp, filename) {
-    // Check if file has been changed first ...
-    if ((this.prevTimestamps[filename] || this.startTime) < (fileTimestamps[filename] || Infinity) &&
-      // ... then validate by the glob pattern.
-      minimatch(filename, glob, {matchBase: true})
-    ) {
-      changedStyleFiles = changedStyleFiles.concat(filename);
-    }
-    return changedStyleFiles;
-  }.bind(this), []);
+LintDirtyModulesPlugin.prototype.getChangedFiles = function getChangedFiles (fileTimestamps, glob) {
+  return Object.keys(fileTimestamps).filter(function (filename) {
+    return hasFileChanged.call(this, filename, fileTimestamps[filename]) && minimatch(filename, glob, {matchBase: true});
+  }.bind(this));
 };
+
+function hasFileChanged (filename, timestamp) {
+  return defaultTo(this.prevTimestamps[filename], this.startTime) < defaultTo(timestamp, Infinity);
+}
 
 module.exports = LintDirtyModulesPlugin;
