+++ package.json
@@ -33,7 +33,7 @@
   "dependencies": {
     "arrify": "^1.0.1",
     "chalk": "^1.1.3",
-    "lodash.reduce": "^4.6.0",
+    "lodash.defaultto": "^4.14.0",
     "minimatch": "^3.0.3",
     "object-assign": "^4.1.0",
     "stylelint": "^7.7.0"
