+++ test/fixtures/lint-free/index.js
