+++ test/fixtures/multiple-sources/index.js
