+++ test/fixtures/single-error/index.js
