--- test/fixtures/test9/index.js
@@ -1,3 +1,0 @@
-require(getPath('./../../../node_modules/file-loader/index') + '!./test.scss');
-
-console.log('test9');
