--- test/fixtures/test9/test.scss
@@ -1,3 +1,0 @@
-body {
-  display: block;
-}
